refactor: spaceInfo 조회 로직 훅 분리
